légére optimisation (4% de rapidité) brutForce v4.1

+++ src/main/java/projetS3Voyageur/StatsAlgos/Analyser.java
@@ -87,8 +87,8 @@ public class Analyser {
     /**
      * @param nbVillesMax the nbVillesMax to set
      */
-    public void setNbVillesMax(byte nbVillesMax) {
-        this.nbVillesMax = nbVillesMax;
+    public void setNbVillesMax(int nbVillesMax) {
+        this.nbVillesMax = (byte) nbVillesMax;
     }
 
     /**
